+++ Model/Import.php
@@ -388,7 +388,7 @@ class Import
         } else {
             // set the time interval
             $this->setIntervalTime(
-                isset($params[self::PARAM_DAYS]) ? (int) $params[self::PARAM_DAYS] : null,
+                isset($params[self::PARAM_DAYS]) ? (float) $params[self::PARAM_DAYS] : null,
                 $params[self::PARAM_CREATED_FROM] ?? null,
                 $params[self::PARAM_CREATED_TO] ?? null
             );
@@ -457,17 +457,25 @@ class Import
     /**
      * Set interval time for order synchronisation
      *
-     * @param integer|null $days Import period
+     * @param float|null $days Import period
      * @param string|null $createdFrom Import of orders since
      * @param string|null $createdTo Import of orders until
      */
-    private function setIntervalTime(int $days = null, string $createdFrom = null, string $createdTo = null): void
+    private function setIntervalTime(float $days = null, string $createdFrom = null, string $createdTo = null): void
     {
 
         if ($createdFrom && $createdTo) {
             // retrieval of orders created from ... until ...
+            if ($createdTo < $createdFrom) {
+                $createdTo = $createdFrom;
+            }
             $createdFromTimestamp = $this->dateTime->gmtTimestamp($createdFrom);
-            $createdToTimestamp = $this->dateTime->gmtTimestamp($createdTo) + 86399;
+            if ($createdFrom === $createdTo) {
+                $createdToTimestamp = $createdFromTimestamp + (self::MIN_INTERVAL_TIME - 1);
+            } else {
+                $createdToTimestamp = $this->dateTime->gmtTimestamp($createdTo);
+            }
+            $createdToTimestamp = $this->dateTime->gmtTimestamp($createdTo);
             $intervalTime = $createdToTimestamp - $createdFromTimestamp;
             $this->createdFrom = $createdFromTimestamp;
             $this->createdTo = $intervalTime > self::MAX_INTERVAL_TIME
@@ -476,12 +484,12 @@ class Import
             return;
         }
         if ($days) {
-            $intervalTime = $days * 86400;
+            $intervalTime = floor($days * self::MIN_INTERVAL_TIME);
             $intervalTime = $intervalTime > self::MAX_INTERVAL_TIME ? self::MAX_INTERVAL_TIME : $intervalTime;
         } else {
             // order recovery updated since ... days
-            $importDays = (int) $this->configHelper->get(ConfigHelper::SYNCHRONIZATION_DAY_INTERVAL);
-            $intervalTime = $importDays * 86400;
+            $importDays = (float) $this->configHelper->get(ConfigHelper::SYNCHRONIZATION_DAY_INTERVAL);
+            $intervalTime = floor($importDays * self::MIN_INTERVAL_TIME);
             // add security for older versions of the plugin
             $intervalTime = $intervalTime < self::MIN_INTERVAL_TIME ? self::MIN_INTERVAL_TIME : $intervalTime;
             $intervalTime = $intervalTime > self::MAX_INTERVAL_TIME ? self::MAX_INTERVAL_TIME : $intervalTime;
@@ -635,8 +643,10 @@ class Import
         }
         try {
             // get orders from Lengow API
-            $orders = $this->getOrdersFromApi($store);
-            $numberOrdersFound = count($orders);
+            $orders = $this->getOrdersFromApi($store, $numberOrdersFound);
+            // current() will trigger the first api call & populate $numberOrdersFound
+            // leaving the cursor at the first element
+            $orders->current();
             if ($this->importOneOrder) {
                 $this->dataHelper->log(
                     DataHelper::CODE_IMPORT,
@@ -751,12 +761,14 @@ class Import
      * Call Lengow order API
      *
      * @param Store $store Magento store instance
+     * @param int|null $count ref . that will be set on the first iteration.
+     *  You can populate the count before iterating by calling $generator->current();
      *
-     * @return array
+     * @return \Generator<array>
      *
-     * @throws LengowException
+     * @throws Exception
      */
-    private function getOrdersFromApi(Store $store): array
+    private function getOrdersFromApi(Store $store, ?int &$count = null): \Generator
     {
         $page = 1;
         $orders = [];
@@ -833,16 +845,20 @@ class Import
                                 ->format(DataHelper::DATE_ISO_8601),
                         ];
                     }
+                    $filterParams = [
+                        self::ARG_CATALOG_IDS => implode(',', $this->storeCatalogIds),
+                        self::ARG_NO_CURRENCY_CONVERSION => $noCurrencyConversion,
+                        self::ARG_ACCOUNT_ID => $this->accountId,
+                        self::ARG_PAGE => $page,
+                    ];
+                    if (!empty($this->marketplaceName)) {
+                        $filterParams[self::ARG_MARKETPLACE] = $this->marketplaceName;
+                    }
                     $results = $this->lengowConnector->get(
                         Connector::API_ORDER,
                         array_merge(
                             $timeParams,
-                            [
-                                self::ARG_CATALOG_IDS => implode(',', $this->storeCatalogIds),
-                                self::ARG_NO_CURRENCY_CONVERSION => $noCurrencyConversion,
-                                self::ARG_ACCOUNT_ID => $this->accountId,
-                                self::ARG_PAGE => $page,
-                            ]
+                            $filterParams
                         ),
                         Connector::FORMAT_STREAM,
                         '',
@@ -879,14 +895,19 @@ class Import
                     )
                 );
             }
+
+            if (null === $count) {
+                $count = $results->count;
+            }
+
             // construct array orders
             foreach ($results->results as $order) {
-                $orders[] = $order;
+                yield $order;
             }
+
             $page++;
             $finish = $results->next === null || $this->importOneOrder;
         } while ($finish !== true);
-        return $orders;
     }
 
     /**