+++ Observer/AddressTotals.php
@@ -69,6 +69,11 @@ class AddressTotals implements ObserverInterface
             return;
         }
 
+        //not fix rounding if total_order, total_tax or shipping is null
+        if (is_null($lengowOrderData->total_order) || is_null($lengowOrderData->total_tax) || is_null($lengowOrderData->shipping)) {
+            return;
+        }
+
         $totalLengow = (float) $lengowOrderData->total_order;
         $taxLengow = (float) $lengowOrderData->total_tax;
         $shippingLengow = (float) $lengowOrderData->shipping;