+++ Plugin/InvoiceTotalsCollectorPlugin.php
@@ -56,6 +56,11 @@ class InvoiceTotalsCollectorPlugin
         }
 
         $lengowOrderData = $this->backendSession->getCurrentOrderLengowData();
+
+        //not fix rounding if total_order, total_tax or shipping is null
+        if (is_null($lengowOrderData->total_order) || is_null($lengowOrderData->total_tax) || is_null($lengowOrderData->shipping)) {
+            return $result;
+        }
         $totalLengow = (float) $lengowOrderData->total_order;
         $taxLengow = (float) $lengowOrderData->total_tax;
         $shippingLengow = (float) $lengowOrderData->shipping;