+++ tools/build.sh
@@ -137,6 +137,8 @@ remove_files $FOLDER_TMP "README.md"
 remove_files $FOLDER_TMP ".gitignore"
 # remove .git
 remove_files $FOLDER_TMP ".git"
+# remove .github
+rm -Rf "$FOLDER_TMP/.github"
 # remove .DS_Store
 remove_files $FOLDER_TMP ".DS_Store"
 # remove .idea